% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebblegraph.R
\name{.pebble}
\alias{.pebble}
\title{Coerce pebble data to raster}
\usage{
.pebble(data, transpose = TRUE)
}
\arguments{
\item{data}{A matrix (or object coercible into type "matrix")}

\item{transpose}{A logical (default \code{TRUE}). If \code{TRUE}, transposes
the input data}
}
\description{
Internal function to coerce pebble data into an object of class
\code{raster}
}
