% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_quantile.R
\name{f_quantile}
\alias{f_quantile}
\title{f_quantile}
\usage{
f_quantile(
  x,
  probs = c(0.05, 0.15, 0.5, 0.84, 0.95),
  names = c("ymin", "lower", "middle", "upper", "ymax"),
  ...
)
}
\arguments{
\item{x}{Numeric Vector}

\item{probs}{Numeric Vector of Probabilities}

\item{names}{Names of probs}

\item{...}{Other arguments to be passed on}
}
\value{
Numeric vector of length(probs)
}
\description{
Helper function to find arbitrary quantiles
}
