% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountaintop.R
\name{mountaintop}
\alias{mountaintop}
\title{Normalize baseline}
\usage{
mountaintop(x, baseline, bottom = TRUE)
}
\arguments{
\item{x}{Numeric Vector}

\item{baseline}{Number or numeric vector to compare against.}

\item{bottom}{Whether or not to "chop off" the bottom or top of the vector.}
}
\value{
Numeric vector
}
\description{
Calculates the parallel minimum (if \code{baseline} is
\code{TRUE}) or parallel maximum, each subtracted by the baseline
measurement.
}
