% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pebble1}
\alias{pebble1}
\title{Sample pebble count data 1}
\format{
A matrix of 20 rows and 8 columns: each column represents a transect
and each cell in that column represents a recording from river left (row 1)
to river right (row 20)
}
\description{
Pebble count data acquired from the Downstream-2 Cross Section 2
site; Sep. 13, 2021
}
\author{
Josh Platt
}
\keyword{datasets}
