% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebblegraph.R
\name{plot_pebble}
\alias{plot_pebble}
\title{Plot Pebble grid as Raster}
\usage{
plot_pebble(data, ..., colNA = "red")
}
\arguments{
\item{data}{A matrix (or object coercible into type "matrix")}

\item{...}{Other arguments to be passed along (primarily to
\code{\link[raster]{raster}})}

\item{colNA}{A character representing a color (eg: "red")}
}
\description{
Plots pebble data (see: \emph{pebble1}) as a raster from
the raster package.
}
