% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{standardize_units}
\alias{standardize_units}
\title{Standardize Units}
\usage{
standardize_units(
  data,
  .cols = tidyselect::eval_select(is.numeric, data),
  Units = NULL
)
}
\arguments{
\item{data}{A data.frame}

\item{.cols}{The columns from which the Units are converted (by default, all numeric).}

\item{Units}{Character of the units of measurement (e.g. "m" or "ft") or default NULL}
}
\value{
Data.frame with all columns in \code{.cols} (by default, all numeric columns) converted to use the units defined in \code{Units}
}
\description{
Converts all columns in a data.frame to use the same system of units
}
