% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot2.R
\docType{data}
\name{StatBoxplot2}
\alias{StatBoxplot2}
\title{Custom geom_boxplot}
\format{
An object of class \code{StatBoxplot2} (inherits from \code{StatBoxplot}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
StatBoxplot2
}
\description{
Creates a boxplot with an arbitrary set of quantiles. Credit to Z.lin on
stackexchange
https://stackoverflow.com/questions/28368963/ggplot-how-to-add-a-segment-with-stat-summary
}
\keyword{datasets}
