% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_sf}
\alias{test_sf}
\title{Test xs in \code{sf} format}
\format{
A data frame with 34 rows and 4 variables:
\describe{
\item{\code{TAPE}}{double Cross-section measurement}
\item{\code{InvertRod}}{double Depth measurement}
\item{\code{Bankful}}{double Baseline measurement}
\item{\code{geometry}}{list \code{geometry} column of class \code{sfc}}
}
}
\description{
Simple example dataset of how to use data created by \code{sf}
}
\author{
Harrison Brown
}
\keyword{datasets}
