% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Sample cross-sectional data}
\format{
A tibble of 34 rows and 3 variables:
\describe{
\item{TAPE}{Tape measurement; x-coordinate}
\item{InvertRod}{Depth measurement in meters}
\item{Bankful}{Bankful measurement}
}
}
\description{
Cross-sectional data acquired from the Watauga River, Boone, N.C., U.S.
}
\author{
Josh Platt

Quincy Williams
}
\keyword{datasets}
