% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_area.R
\name{xs_area}
\alias{xs_area}
\title{Calculate cross-sectional area}
\usage{
xs_area(x)
}
\arguments{
\item{x}{Object of class \code{xs}}
}
\value{
Numeric indicating cross-sectional area, or object of class
\code{"units"} if the \code{xs} object has units.
}
\description{
This function calculates the area of a cross section object
(\code{xs})
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){

 }
}
}
