% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trap_area}
\alias{trap_area}
\title{Calculate trapezoidal-rule area for two vectors}
\usage{
trap_area(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
A double of the calculated area
}
\description{
Calculate trapezoidal-rule area for two vectors
}
\keyword{internal}
