% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_area.R
\name{xs_area}
\alias{xs_area}
\title{Cross-Sectional area}
\usage{
xs_area(data, tape, depth, baseline, sum_area = TRUE)
}
\arguments{
\item{data}{A data.frame of cross-section data}

\item{tape}{A column from \code{data} or a numeric vector of tape readings
("x coordinates").}

\item{depth}{A column from \code{data} or a numeric vector of rod readings
("y coordinates").}

\item{baseline}{A column from \code{data} or vector representing bankfull
level.}

\item{sum_area}{A logical (default \code{TRUE}) indicating whether to return
a single summary of the trapezoidal area or a tibble indicating the area at
each section.}
}
\value{
A Double of the cross-sectional area compared to a baseline
}
\description{
Calculate area of a cross-section using the trapezoidal rule for
integration. By default, uses the columns "TAPE" for measurement across xs,
"InvertRod" for depth relative to y = 0, and "Bankful" for baseline
measurement.
}
