% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_area.R
\name{xs_area}
\alias{xs_area}
\title{Calculate Cross-sectional area}
\usage{
xs_area(.data, x_cor = TAPE, invertrod = InvertRod, baseline = Bankful)
}
\arguments{
\item{.data}{A data.frame of cross-section data}

\item{x_cor}{A column of x-coordinates (default, \emph{TAPE})}

\item{invertrod}{A column of rod readings, relative to 0 (i.e. >0)}

\item{baseline}{A column of baseline readings, defaults to Bankful}
}
\value{
A Double of the cross-sectional area compared to a baseline
}
\description{
Calculate Cross-sectional area
}
