% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_area.R
\name{xs_area}
\alias{xs_area}
\title{Cross-Sectional area}
\usage{
xs_area(data, Units = NULL, tape = NULL, depth = NULL, baseline = NULL)
}
\arguments{
\item{data}{A data.frame of cross-section data}

\item{tape}{Either a numeric vector of tape readings ("x coordinates").}

\item{depth}{Either a numeric vector of rod readings ("y coordinates").}

\item{baseline}{Either a double representing bankfull level.}
}
\value{
A Double of the cross-sectional area compared to a baseline
}
\description{
Calculate area of a cross-section using the trapezoidal rule for
integration. By default, uses the columns "TAPE" for measurement across xs,
"InvertRod" for depth relative to y = 0, and "Bankful" for baseline
measurement.
}
\examples{
area <- xs_area(testxs)
print(area)

}
