% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_area.R
\name{xs_area}
\alias{xs_area}
\title{Calculate XS area from data}
\usage{
xs_area(.data)
}
\arguments{
\item{.data}{A data.frame of cross-section data}
}
\value{
A Double of the cross-sectional area compared to a baseline (default, \emph{Bankful})
}
\description{
Calculate XS area from data
}
