% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_ribbon.R
\name{xs_ribbon}
\alias{xs_ribbon}
\title{Generate an area plot in ggplot2}
\usage{
xs_ribbon(.data)
}
\arguments{
\item{.data}{Cross section data}
}
\value{
A ggplot
}
\description{
Generate an area plot in ggplot2
}
